Added Various Scale Transforms
This change adds several functions for transforming scales, and changes the
behavior of the tonic property of the Scale class to be useful.

The tonic property can be set from note names like "C#" or as an integer offset
(in half steps) above C.  When read, the tonic property will always return the
integer offset (in half steps) of the note above C.

The following transforms have been added:

 - rename_to_matching_mode

 - sharpen

 - flatten

 - rotate

 - clink_my_heptatonic

The docstrings for these functions explain their intended use.