Initial Implementation of Chord and Interval Classes
This change pulls the code for quantifying intervals out of analysis.py and puts
it in the new chords.py file.  This file also contains the new Interval and Chord
classes.

The Interval class takes two notes and an optional expected interval number or enum
as constructor arguments, and the class provides the actual interval enum as well
as the quality of the interval.

The Chord class takes a root note and some number of intervals as constructor
arguments, and the intervals must all have the same root which the chord was
constructed with.

The Chord class will eventually be used for determining the name of the chord, and
in other things like scale analysis.

The way the chord class is constructed right now is gnarly, but I intend to add
helper functions in the future for constructing chords from, say, a bag of notes.