New Chord System
This tears out the chords / intervals system I had before.  From discussion on
The Internet, it appears that the standard chord names are arbitrary, and that
there is no way to enumerate them automatically.  So we have a lookup table now.

The new Chord class works kind of like the Scale class - both internally are
represented as a "keyboard" tuple.  The specifics are a little different, but
I imagine these could be merged together at some point.

Anyways, this now has the ability to enumerate every chord in a scale.
Since the result is a set, diffing the chords between two scales is super easy now.